Move context to HttpServer, HttpApp via inheritence
